Bookbuddy_app/requirements.txt
Python 3.12.4
Flask 3.0.3
Werkzeug 3.0.4
Bootstrap 5.3.0
request in Python installed with pip
flask-wtf email-validator
flask-talisman
install Flask-Limiter==3.3.0
bleach
pytest
pytest-cov
google-api-python-client cachetools
